requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
asyncpg==0.30.0
python-jose==3.3.0
passlib[argon2]==1.7.4
python-multipart==0.0.20
pydantic-settings==2.7.0
email-validator==2.1.1
cloudinary==1.41.0
python-dotenv==1.0.1
